Parse Investigation ID from JSON

+++ src/eaglecore_models.rs
@@ -1,5 +1,5 @@
 pub struct Investigation {
-    pub id: i16,
+    pub id: u64,
     pub name: String
 }
